+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Remove Rigged Faire Game secret when opponent loses armor
 - Fix memory growth due to spectator improvements (issue #1207)
 - Fix secret helper in arena (#issue 1222)
+- Exclude BG neutral minions that are tribe specific
 
 # 1.9.4
 ## New